requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
faster-whisper
groq
jinja2
pdfkit
python-dotenv
reportlab